progress on Material management between thread
