+++ GameEngine/GameEngine/Engine/Threads/PrepareRenderThread.cpp
@@ -106,6 +106,38 @@ namespace AGE
 			GetRenderThread()->stop();
 		});
 
+		registerCallback<Tasks::MainToPrepare::AddMaterial>([&](Tasks::MainToPrepare::AddMaterial &msg)
+		{
+			assert(this->_activeScene != nullptr);
+			_activeScene->_addMaterial(msg);
+			
+		});
+
+		registerCallback<Tasks::MainToPrepare::SetMeshMaterial>([&](Tasks::MainToPrepare::SetMeshMaterial& msg)
+		{
+			//			for (auto &subMesh : msg.mesh->subMeshs)
+			//			{
+			//				auto vertices = paintingManager->get_painter(subMesh.painter)->get_vertices(subMesh.vertices);
+			//				assert(vertices != nullptr);
+			//				if (subMesh.defaultMaterialIndex >= msg.material->datas.size())
+			//				{
+			//					for (auto &prop : msg.material->datas[0])
+			//					{
+			//						prop->set_program(pipelines[0]->get_programs());
+			//						vertices->add_property(prop);
+			//					}
+			//				}
+			//				else
+			//				{
+			//					for (auto &prop : msg.material->datas[subMesh.defaultMaterialIndex])
+			//					{
+			//						prop->set_program(pipelines[0]->get_programs());
+			//						vertices->add_property(prop);
+			//					}
+			//				}
+			//			}
+		});
+
 		return true;
 	}
