+++ GameEngine/GameEngine/Engine/Threads/RenderThread.cpp
@@ -19,18 +19,7 @@
 #include <SpacePartitioning/Ouptut/RenderCamera.hh>
 #include <SpacePartitioning/Ouptut/RenderLight.hh>
 #include <SpacePartitioning/Ouptut/RenderPipeline.hh>
-#include <Render/Properties/Materials/Color.hh>
-#include <Render/Properties/Materials/MapColor.hh>
 #include <Utils/Debug.hpp>
-
-#define CREATE_MATERIAL_COLOR(memberName, uniformName)	tmpColor = std::make_shared<Color>(std::string(uniformName)); \
-	tmpColor->set(msg.data.memberName); \
-	instance._properties[uniformName] = properties.add_property(tmpColor);
-
-#define CREATE_MATERIAL_MAP(memberName, uniformName)	if (!msg.data.memberName.empty()) { \
-		tmpMap = std::make_shared<MapColor>(std::string(uniformName)); \
-		instance._properties[uniformName] = properties.add_property(tmpMap); \
-	}
 
 namespace AGE
 {
@@ -66,33 +55,6 @@ namespace AGE
 			glClear(GL_COLOR_BUFFER_BIT);
 		});
 
-		registerCallback<Tasks::Render::AddMaterial>([&](Tasks::Render::AddMaterial &msg)
-		{
-			MaterialInstance instance;
-			_materials.emplace_back();
-			Material &properties = _materials.back();
-
-			instance._material_key = Key<Material>::createKey(_materials.size() - 1);
-
-			std::shared_ptr<Color> tmpColor = nullptr;
-			std::shared_ptr<MapColor> tmpMap = nullptr;
-
-			CREATE_MATERIAL_COLOR(diffuse, "diffuseColor");
-			CREATE_MATERIAL_COLOR(ambient, "ambientColor");
-			CREATE_MATERIAL_COLOR(emissive, "emissiveColor");
-			CREATE_MATERIAL_COLOR(reflective, "reflectiveColor");
-			CREATE_MATERIAL_COLOR(specular, "specularColor");
-
-			CREATE_MATERIAL_MAP(diffuseTexPath, "diffuseMap");
-			CREATE_MATERIAL_MAP(ambientTexPath, "ambientMap");
-			CREATE_MATERIAL_MAP(emissiveTexPath, "emissiveMap");
-			CREATE_MATERIAL_MAP(reflectiveTexPath, "reflectiveMap");
-			CREATE_MATERIAL_MAP(specularTexPath, "specularMap");
-			CREATE_MATERIAL_MAP(normalTexPath, "normalMap");
-			CREATE_MATERIAL_MAP(bumpTexPath, "bumpMap");
-
-			msg.setValue(instance);
-		});
 
 		registerCallback<Tasks::Render::GetWindowSize>([&](Tasks::Render::GetWindowSize &msg)
 		{
@@ -154,31 +116,6 @@ namespace AGE
 			AGE::Imgui::getInstance()->renderThreadRenderFn(msg.cmd_lists);
 		});
 #endif
-
-		registerCallback<Tasks::Render::SetMeshMaterial>([&](Tasks::Render::SetMeshMaterial& msg)
-		{
-//			for (auto &subMesh : msg.mesh->subMeshs)
-//			{
-//				auto vertices = paintingManager->get_painter(subMesh.painter)->get_vertices(subMesh.vertices);
-//				assert(vertices != nullptr);
-//				if (subMesh.defaultMaterialIndex >= msg.material->datas.size())
-//				{
-//					for (auto &prop : msg.material->datas[0])
-//					{
-//						prop->set_program(pipelines[0]->get_programs());
-//						vertices->add_property(prop);
-//					}
-//				}
-//				else
-//				{
-//					for (auto &prop : msg.material->datas[subMesh.defaultMaterialIndex])
-//					{
-//						prop->set_program(pipelines[0]->get_programs());
-//						vertices->add_property(prop);
-//					}
-//				}
-//			}
-		});
 
 		return true;
 	}
