+++ GameEngine/GameEngine/Engine/Threads/Tasks/MainToPrepareTasks.hpp
@@ -2,6 +2,7 @@
 #define __TASKS_MAIN_TO_PREPARE__
 
 #include <TMQ/message.hpp>
+#include <AssetManagement/Data/MaterialData.hh>
 
 namespace AGE
 {
@@ -15,7 +16,28 @@ namespace AGE
 				CreateScene(AScene *_scene);
 				AScene *scene;
 			};
-		};
+
+			struct AddMaterial : TMQ::FutureData < MaterialInstance >
+			{
+				MaterialData data;
+				AddMaterial(MaterialData const &pData) :
+					data(pData) {}
+			};
+
+			struct SetMeshMaterial
+			{
+				std::shared_ptr<MaterialSetInstance> material;
+				std::shared_ptr<MeshInstance> mesh;
+
+				SetMeshMaterial(std::shared_ptr<MaterialSetInstance> pMaterial,
+					std::shared_ptr<MeshInstance> pMesh) :
+					material(pMaterial),
+					mesh(pMesh)
+				{ }
+
+			};
+		}
+
 	}
 }
 #endif
